# docs/development/api/swaggerStuff.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Tutor Auth Controller","description":"Tutor Authentication and Authorization Endpoints"},{"name":"Catalog Controller","description":"Catalog of Courses"},{"name":"Intermediary Payment Controller","description":"Intermediary controller handling payment requests"},{"name":"Tutor Payment Controller","description":"Tutor sends payment requests"},{"name":"Tutor Controller","description":"Tutor profile management endpoints"},{"name":"Matchmaking Controller","description":"Student-tutor matchmaking endpoints"},{"name":"User Controller","description":"User profile management endpoints"},{"name":"Student Payment Controller","description":"Student accepts or rejects payment requests"},{"name":"User Auth Controller","description":"User Authentication and Authorization Endpoints"}],"paths":{"/users/{userId}/update":{"put":{"tags":["User Controller"],"summary":"Modify user profile","description":"Update user profile","operationId":"updateUserProfile","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserProfileReq"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserProfileResp"}}}}}}},"/tutors/{tutorId}/subjects":{"put":{"tags":["Tutor Controller"],"summary":"Select or modify tutor subjects","description":"Manage tutor subjects","operationId":"manageSubjects","parameters":[{"name":"tutorId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TutorSubjectsReq"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/tutors/{tutorId}/profile":{"get":{"tags":["Tutor Controller"],"summary":"Retrieve tutor profile","description":"Get tutor profile","operationId":"getTutorProfile","parameters":[{"name":"tutorId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TutorProfileResp"}}}}}},"put":{"tags":["Tutor Controller"],"summary":"Modify tutor profile","description":"Update tutor profile","operationId":"updateTutorProfile","parameters":[{"name":"tutorId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTutorProfileReq"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TutorProfileResp"}}}}}}},"/users/{userId}/upgrade":{"post":{"tags":["User Controller"],"summary":"Upgrade user to gold","description":"Upgrade to gold user","operationId":"upgradeToGold","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/user/auth/register":{"post":{"tags":["User Auth Controller"],"summary":"User registration","description":"Register user","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegistrationReq"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserInfoResp"}}}}}}},"/user/auth/login":{"post":{"tags":["User Auth Controller"],"summary":"User login","description":"Login user","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCredentialsReq"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserInfoResp"}}}}}}},"/tutor/auth/register":{"post":{"tags":["Tutor Auth Controller"],"summary":"Tutor registration","description":"Register tutor","operationId":"register_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegistrationReq"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserInfoResp"}}}}}}},"/tutor/auth/login":{"post":{"tags":["Tutor Auth Controller"],"summary":"Tutor login","description":"Login tutor","operationId":"login_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCredentialsReq"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserInfoResp"}}}}}}},"/payments/tutor/request":{"post":{"tags":["Tutor Payment Controller"],"summary":"Send payment request","description":"Tutor sends a payment request to the intermediary controller","operationId":"sendPaymentRequest","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaymentReq"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaymentResp"}}}}}}},"/payments/student/{paymentRequestId}/reject":{"post":{"tags":["Student Payment Controller"],"summary":"Reject payment request","description":"Student rejects the payment request","operationId":"rejectPayment","parameters":[{"name":"paymentRequestId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaymentResp"}}}}}}},"/payments/student/{paymentRequestId}/accept":{"post":{"tags":["Student Payment Controller"],"summary":"Accept payment request","description":"Student accepts the payment request","operationId":"acceptPayment","parameters":[{"name":"paymentRequestId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AcceptPaymentReq"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaymentResp"}}}}}}},"/payments/intermediary/forward":{"post":{"tags":["Intermediary Payment Controller"],"summary":"Forward payment request","description":"Intermediary forwards the payment request to the student controller","operationId":"forwardPaymentRequest","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaymentReq"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaymentResp"}}}}}}},"/matchmaking/validate-tutor-authentication":{"post":{"tags":["Matchmaking Controller"],"summary":"Validate tutor authentication","description":"Validate tutor authentication","operationId":"validateTutorAuthentication","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidateAuthReq"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/matchmaking/validate-student-authentication":{"post":{"tags":["Matchmaking Controller"],"summary":"Validate student authentication","description":"Validate student authentication","operationId":"validateStudentAuthentication","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidateAuthReq"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/matchmaking/request":{"post":{"tags":["Matchmaking Controller"],"summary":"Send matchmaking request","description":"Send match request","operationId":"sendMatchRequest","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MatchRequestReq"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/api/subscription/{studentId}":{"get":{"tags":["subscription-controller"],"operationId":"getSubscriptionStatus","parameters":[{"name":"studentId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SubscriptionResp"}}}}}},"post":{"tags":["subscription-controller"],"operationId":"subscribe","parameters":[{"name":"studentId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubscriptionReq"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SubscriptionResp"}}}}}}},"/api/send-message":{"post":{"tags":["messaging-controller"],"operationId":"sendMessage","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/ratings":{"post":{"tags":["rating-controller"],"operationId":"submitRating","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RatingDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/notifications":{"post":{"tags":["notification-controller"],"operationId":"createNotification","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotificationReq"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NotificationResp"}}}}}}},"/users/{userId}/profile":{"get":{"tags":["User Controller"],"summary":"Retrieve user profile details","description":"Get user profile","operationId":"getUserProfile","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserProfileResp"}}}}}}},"/catalog/courses":{"get":{"tags":["Catalog Controller"],"summary":"View Course Catalog","description":"Get all available courses","operationId":"getAllCourses","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CourseResp"}}}}}}}},"/catalog/courses/search":{"get":{"tags":["Catalog Controller"],"summary":"Search Courses","description":"Search courses by subject","operationId":"searchCourses","parameters":[{"name":"subject","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CourseResp"}}}}}}}},"/api/ratings/{tutorId}":{"get":{"tags":["rating-controller"],"operationId":"getRatingsByTutor","parameters":[{"name":"tutorId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RatingDTO"}}}}}}}},"/api/notifications/{notificationId}":{"get":{"tags":["notification-controller"],"operationId":"getNotificationById","parameters":[{"name":"notificationId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NotificationResp"}}}}}}},"/api/notifications/user/{userId}":{"get":{"tags":["notification-controller"],"operationId":"getNotificationsByUser","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/NotificationResp"}}}}}}}},"/users/{userId}/delete":{"delete":{"tags":["User Controller"],"summary":"Delete user profile","description":"Delete user account","operationId":"deleteUser","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/tutors/{tutorId}/delete":{"delete":{"tags":["Tutor Controller"],"summary":"Delete tutor account","description":"Delete tutor profile","operationId":"deleteTutor","parameters":[{"name":"tutorId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}}},"components":{"schemas":{"UpdateUserProfileReq":{"type":"object","properties":{"fullName":{"type":"string"},"email":{"type":"string"},"role":{"type":"string"}}},"UserProfileResp":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"fullName":{"type":"string"},"email":{"type":"string"},"role":{"type":"string"}}},"TutorSubjectsReq":{"type":"object","properties":{"preMadeSubjects":{"type":"array","items":{"type":"string"}},"customSubjects":{"type":"array","items":{"type":"string"}}}},"UpdateTutorProfileReq":{"type":"object","properties":{"fullName":{"type":"string"},"subjects":{"type":"array","items":{"type":"string"}},"hourlyRates":{"type":"object","additionalProperties":{"type":"number","format":"double"}}}},"TutorProfileResp":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"fullName":{"type":"string"},"subjects":{"type":"array","items":{"type":"string"}},"hourlyRates":{"type":"object","additionalProperties":{"type":"number","format":"double"}}}},"UserRegistrationReq":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"},"email":{"type":"string"},"fullName":{"type":"string"}}},"UserInfoResp":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"username":{"type":"string"},"type":{"type":"string"}}},"UserCredentialsReq":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"PaymentReq":{"type":"object","properties":{"sessionId":{"type":"integer","format":"int64"},"studentId":{"type":"integer","format":"int64"},"tutorId":{"type":"integer","format":"int64"},"amount":{"type":"number","format":"double"},"subject":{"type":"string"},"hourlyRate":{"type":"number","format":"double"},"duration":{"type":"integer","format":"int32"}}},"PaymentResp":{"type":"object","properties":{"paymentRequestId":{"type":"integer","format":"int64"},"message":{"type":"string"},"transactionId":{"type":"string"},"status":{"type":"string"}}},"AcceptPaymentReq":{"type":"object","properties":{"studentId":{"type":"integer","format":"int64"},"paymentRequestId":{"type":"integer","format":"int64"}}},"ValidateAuthReq":{"type":"object","properties":{"token":{"type":"string"}}},"MatchRequestReq":{"type":"object","properties":{"studentToken":{"type":"string"},"tutorId":{"type":"integer","format":"int64"}}},"SubscriptionReq":{"type":"object","properties":{"subscriptionType":{"type":"string"},"paymentMethod":{"type":"string"},"durationMonths":{"type":"integer","format":"int32"}}},"SubscriptionResp":{"type":"object","properties":{"studentId":{"type":"integer","format":"int64"},"subscriptionType":{"type":"string"},"startDate":{"type":"string","format":"date"},"endDate":{"type":"string","format":"date"},"status":{"type":"string"}}},"MessageDTO":{"type":"object","properties":{"message":{"type":"string"}}},"RatingDTO":{"type":"object","properties":{"userId":{"type":"integer","format":"int64"},"tutorId":{"type":"integer","format":"int64"},"rating":{"type":"integer","format":"int32"},"comment":{"type":"string"}}},"NotificationReq":{"type":"object","properties":{"userId":{"type":"integer","format":"int64"},"message":{"type":"string"},"type":{"type":"string"}}},"NotificationResp":{"type":"object","properties":{"notificationId":{"type":"integer","format":"int64"},"userId":{"type":"integer","format":"int64"},"message":{"type":"string"},"type":{"type":"string"},"timestamp":{"type":"string","format":"date-time"}}},"CourseResp":{"type":"object","properties":{"courseId":{"type":"integer","format":"int64"},"courseName":{"type":"string"},"tutorName":{"type":"string"},"hourlyRate":{"type":"number","format":"double"},"rating":{"type":"number","format":"double"}}}}}}